Split executable into kabanero-events and kabanero-listener.
