+++ Dockerfile.events
@@ -38,7 +38,7 @@ RUN go get -u golang.org/x/lint/golint; golint -set_exit_status
 # RUN go test -v
 
 # Build executable
-RUN CGO_ENABLED=0 GOOS=linux go build -a -ldflags '-extldflags "-static"'
+RUN CGO_ENABLED=0 GOOS=linux go build -a -ldflags '-extldflags "-static"' github.com/kabanero-io/kabanero-events/cmd/kabanero-events/...
 
 # Stage 2: Build official image based on UBI
 FROM registry.access.redhat.com/ubi7-minimal:7.6-123
@@ -68,5 +68,5 @@ WORKDIR /app
 EXPOSE 9443
 
 # run with log level 2
-# Note liveness/readiness probe depends on './events'
+# Note liveness/readiness probe depends on './kabanero-events'
 CMD ./kabanero-events -v 5
