+++ pkg/messages/nats_provider.go
@@ -1,4 +1,4 @@
-package main
+package messages
 
 import (
 	"fmt"
