+++ pkg/messages/rest_provider.go
@@ -1,4 +1,4 @@
-package main
+package messages
 
 import (
 	"bytes"
